+++ sqlite3/assets/wasm/bridge.h
@@ -54,6 +54,8 @@ import_dart("function_hook") extern void dartUpdateHook(void *id, int kind,
                                                         const char *db,
                                                         const char *table,
                                                         sqlite3_int64 rowid);
+import_dart("function_commit_hook") extern int dartCommitHook(void *id);
+import_dart("function_rollback_hook") extern void dartRollbackHook(void *id);
 import_dart("function_compare") extern int dartXCompare(void *id, int lengthA,
                                                         const void *a,
                                                         int lengthB,