+++ sqlite3/assets/wasm/helpers.c
@@ -227,6 +227,14 @@ SQLITE_API void dart_sqlite3_updates(sqlite3 *db, int id) {
   sqlite3_update_hook(db, id >= 0 ? &dartUpdateHook : NULL, (void *)id);
 }
 
+SQLITE_API void dart_sqlite3_commits(sqlite3 *db, int id) {
+  sqlite3_commit_hook(db, id >= 0 ? &dartCommitHook : NULL, (void *)id);
+}
+
+SQLITE_API void dart_sqlite3_rollbacks(sqlite3 *db, int id) {
+  sqlite3_rollback_hook(db, id >= 0 ? &dartRollbackHook : NULL, (void *)id);
+}
+
 SQLITE_API int dart_sqlite3_create_collation(sqlite3 *db, const char *zName,
                                              int eTextRep, int id) {
   return sqlite3_create_collation_v2(db, zName, eTextRep, (void *)id,